Inject new lifecycle aware MeshServiceClient
